Consolidate database scope services to single IDatabaseScopeService
- Merge IScopedDatabaseService functionality into IDatabaseScopeService interface
- Update DatabaseScopeService to include scoped caching and transaction coordination
- Remove IScopedDatabaseService and ScopedDatabaseService redundant files
- Register IDatabaseScopeService as scoped service for per-request coordination
- Update all tests and services to use consolidated interface
- Rename ScopedDatabaseServiceTests to DatabaseScopeServiceTests
- All memory and database tests still passing (14/14) with simplified architecture

Eliminates interface duplication while maintaining scoped transaction coordination.
